[#951] Move vector storage class to its own header
